requirements.txt
# Core dependencies
pandas>=2.0.0
duckdb>=1.2.1
numpy>=1.24.0
pyyaml>=6.0.1
rich>=13.7.0

# Visualization dependencies
matplotlib>=3.7.0
seaborn>=0.12.0

# Schema visualization
networkx>=3.0

# Interactive SQL
prompt_toolkit>=3.0.33
pygments>=2.15.0

# Table formatting
tabulate>=0.9.0

# API libraries (from original requirements)
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.1
typer>=0.9.0
mplfinance>=0.12.10b0
python-dateutil>=2.8.2
pytz>=2024.1
beautifulsoup4>=4.12.0
pandas-market-calendars>=3.0